Fixed buffer underrun on invalid configuration.
If a configuration file contains a line which starts
with an equal sign (=), xrr-events triggers an out
of boundary access.

The code in questions assumes that there will be at
least one character before the equal sign and starts
scanning for whitespace characters.

The illegal access can be seen when xrr-events is
compiled with ASAN.

$ echo = > $HOME/.config/xrr-events/xrr-events.conf

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>